+++ packages/common-test-utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.1-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/common-test-utils@3.0.0...@ceramicnetwork/common-test-utils@3.0.1-rc.0) (2024-02-08)
+
+**Note:** Version bump only for package @ceramicnetwork/common-test-utils
+
+
+
+
+
+# [3.0.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/common-test-utils@2.1.0...@ceramicnetwork/common-test-utils@3.0.0) (2024-02-08)
+
+**Note:** Version bump only for package @ceramicnetwork/common-test-utils
+
+
+
+
+
 # [2.1.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/common-test-utils@2.1.0-rc.0...@ceramicnetwork/common-test-utils@2.1.0) (2024-02-02)
 
 **Note:** Version bump only for package @ceramicnetwork/common-test-utils