+++ packages/pinning-ipfs-backend/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.1-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-ipfs-backend@5.0.0...@ceramicnetwork/pinning-ipfs-backend@5.0.1-rc.0) (2024-02-08)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-ipfs-backend
+
+
+
+
+
+# [5.0.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-ipfs-backend@4.1.0...@ceramicnetwork/pinning-ipfs-backend@5.0.0) (2024-02-08)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-ipfs-backend
+
+
+
+
+
 # [4.1.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/pinning-ipfs-backend@4.1.0-rc.0...@ceramicnetwork/pinning-ipfs-backend@4.1.0) (2024-02-02)
 
 **Note:** Version bump only for package @ceramicnetwork/pinning-ipfs-backend